server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
llama-index==0.9.15
llama-index-embeddings-openai==0.1.7
weaviate-client==3.25.3
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
pdfplumber==0.10.3
PyMuPDF==1.23.8
aiofiles==23.2.1
